fix(payment): CodeQL指摘事項を修正 (#110)
- MockPaymentService: 整数乗算のオーバーフロー防止（先にdoubleに変換）
- SupabasePaymentService: 使用されないrequest変数を削除

+++ Baketa.Infrastructure/Payment/Services/MockPaymentService.cs
@@ -200,8 +200,8 @@ public sealed class MockPaymentService : IPaymentService
 
         if (cycle == BillingCycle.Yearly)
         {
-            // 年額は20%OFF
-            return (int)(monthlyPrice * 12 * 0.8);
+            // 年額は20%OFF（オーバーフロー防止のため先にdoubleに変換）
+            return (int)((double)monthlyPrice * 12 * 0.8);
         }
 
         return monthlyPrice;
